requirements.txt
asgiref==3.4.1
black==21.12b0
click==8.0.3
Django==3.2.11
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
importlib-metadata==4.10.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
PyJWT==2.1.0
pytz==2021.3
sqlparse==0.4.2
tomli==1.2.3
typed-ast==1.5.1
typing-extensions==4.0.1
zipp==3.7.0